+++ rpcs3/Emu/RSX/VK/VKGSRender.h
@@ -53,8 +53,6 @@ extern u64 get_system_time();
 
 namespace vk
 {
-	struct buffer_view;
-
 	struct command_buffer_chunk: public vk::command_buffer
 	{
 		vk::fence* submit_fence = nullptr;
