+++ rpcs3/Emu/RSX/VK/VKPresent.cpp
@@ -1,6 +1,5 @@
 #include "stdafx.h"
 #include "VKGSRender.h"
-#include "vkutils/buffer_view.h"
 #include "Emu/Cell/Modules/cellVideoOut.h"
 
 #include "util/asm.hpp"
