+++ rpcs3/Emu/RSX/VK/VKVertexBuffers.cpp
@@ -2,7 +2,6 @@
 #include "VKGSRender.h"
 #include "../Common/BufferUtils.h"
 #include "../rsx_methods.h"
-#include "vkutils/buffer_view.h"
 
 namespace vk
 {
