+++ rpcs3/VKGSRender.vcxproj.filters
@@ -44,15 +44,6 @@
     <ClCompile Include="Emu\RSX\VK\vkutils\shared.cpp">
       <Filter>vkutils</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\vkutils\buffer_view.cpp">
-      <Filter>vkutils</Filter>
-    </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\vkutils\render_device.cpp">
-      <Filter>vkutils</Filter>
-    </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\vkutils\command_pool.cpp">
-      <Filter>vkutils</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Emu\RSX\VK\VKCommonDecompiler.h" />
@@ -109,15 +100,6 @@
       <Filter>vkutils</Filter>
     </ClInclude>
     <ClInclude Include="Emu\RSX\VK\vkutils\supported_extensions.h">
-      <Filter>vkutils</Filter>
-    </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\vkutils\buffer_view.h">
-      <Filter>vkutils</Filter>
-    </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\vkutils\render_device.h">
-      <Filter>vkutils</Filter>
-    </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\vkutils\command_pool.h">
       <Filter>vkutils</Filter>
     </ClInclude>
   </ItemGroup>
